[Chore #168849504] Restructured queries in project
