+++ api/src/auth/products.js
@@ -2,7 +2,7 @@
 import protocol from '../helpers/response';
 import database from '../db/pgConnect';
 import Errors from '../helpers/errors';
-import Queries from '../helpers/queries';
+import Queries from '../queries/products';
 
 export default class Products {
   static async verifyProductName({ body }, res, next) {
