+++ doc/products/createOne.yaml
@@ -11,13 +11,13 @@ paths:
          content:
           application/json:
             schema:
-              $ref: '#/components/schemas/Product'
+              $ref: '#/components/schemas/AddProduct'
       responses:          # server responses
         201:
           description: An object with product details
 components:        # Schema definition for request body
  schemas:
-  Bus:
+  Add Product:
     type: object
     properties:
           name:
